requirements.txt
discord.py
gspread
google-generativeai